refactor: Eliminate duplicate logic and clean up code
