--- src/main/java/event/nbc/event/service/S3Service.java
@@ -1,32 +1,0 @@
-package event.nbc.event.service;
-
-
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.S3Object;
-import com.amazonaws.services.s3.model.S3ObjectInputStream;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-
-import java.io.IOException;
-
-@Service
-@RequiredArgsConstructor
-public class S3Service {
-
-    private final AmazonS3 s3Client;
-
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucket;
-
-    public byte[] getImageBytes(String key) {
-        S3Object s3Object = s3Client.getObject(bucket, key);
-
-        try (S3ObjectInputStream inputStream = s3Object.getObjectContent()) {
-            return inputStream.readAllBytes();
-        } catch (IOException e) {
-            throw new RuntimeException("이미지 읽기 실패", e);
-        }
-    }
-}
-
