+++ src/main/java/event/nbc/s3/S3config.java
@@ -45,4 +45,5 @@ public class S3config {
 			.credentialsProvider(StaticCredentialsProvider.create(credentials))
 			.build();
 	}
+
 }
