--- src/main/resources/s3.yml
@@ -1,9 +1,0 @@
-cloud:
-  aws:
-    credentials:
-      access-key: ${s3.access.key}
-      secret-key: ${s3.secret.key}
-    region:
-      static: ap-northeast-2
-    s3:
-      bucket: ${s3.bucket-name}
